+++ src/Oro/Bundle/DataGridBundle/Resources/public/js/datagrid/grid.js
@@ -647,6 +647,7 @@ define(function(require) {
                 // Skip to add specific attributes if this cell has an empty value.
                 notMarkAsBlank: true
             });
+
             return column;
         },